refactor/환경 수정
