+++ .prettierrc.cjs
@@ -1,0 +1,9 @@
+/**
+ * @type { import('prettier').Config }
+ */
+const config = {
+  printWidth: 80,
+  semi: true,
+};
+
+module.exports = config;
