+++ src/App.tsx
@@ -1,0 +1,6 @@
+import "./App.css";
+import Binspector from "./Pages/Binspector/Binspector";
+
+export function App() {
+  return <Binspector />;
+}
