+++ src/Components/README.md
@@ -1,0 +1,5 @@
+도메인 없는 리액트 컴포넌트 디렉토리
+
+- 외부 디자인 시스템을 사용하지만, Atomic Design 디렉토리 구조로 필요한 부분 있으면 추가
+
+- 컴포넌트를 정의한 파일 및 디렉토리 네이밍은 대문자 시작.
