+++ src/Pages/Binspector/EditorPanel/EditorPanel.css
