+++ src/Pages/Binspector/EditorPanel/StringPanel.tsx
@@ -1,11 +1,7 @@
-import "./Editor.css";
-
-function StringPanel() {
+export function StringPanel() {
   return (
     <div className="string-panel">
       <h1>StringPanel</h1>
     </div>
   );
 }
-
-export default StringPanel;
