+++ src/Pages/Binspector/InfoPanel/InfoPanel.tsx
@@ -1,11 +1,7 @@
-import "../../Binspector.css";
-
-function InfoPanel() {
+export function InfoPanel() {
   return (
     <div className="panel portion-info">
       <h2>Info Panel</h2>
     </div>
   );
 }
-
-export default InfoPanel;
