+++ src/hooks/README.md
@@ -1,0 +1,4 @@
+도메인 디펜던시 없는 리액트 훅, 유틸 함수 관리 디렉토리
+
+- 리액트 훅: use로 시작하는 함수 네이밍 컨벤션
+- useState/useEffect/useRef 리액트 의존하면, Pure Function이 될 수 없음.
