+++ src/utils/README.md
@@ -1,0 +1,4 @@
+아래 디펜던시 없는 제작 모듈 관리 ( 외부 모듈 래퍼는 분리해도 됨.)
+
+- 도메인 디펜던시
+- SPA 라이브러리 디펜던시
